Moved PanelReinforcement objects to ...SurfaceProperties.Reinforcement namespace and updated descriptions
